+++ mypetcare/src/main/webapp/WEB-INF/views/service/Q&A.jsp
@@ -1,0 +1,14 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+	pageEncoding="UTF-8"%>
+<%@ include file="../header.jsp"%>
+<%@ include file="serviceMenu.jsp"%>
+
+
+
+
+<body>
+	<div align="center"><h1>Q&A</h1></div><br><br><br><br>
+</body>
+
+
+<%@ include file="footer.jsp"%>
