+++ mypetcare/src/main/webapp/WEB-INF/views/service/serviceMenu.jsp
@@ -1,0 +1,9 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+
+
+<div class=serviceMenu>
+<h1>고객센터</h1>
+	<div><a onClick="location.href='/service/notice'">공지사항</a></div>
+	<div><a onClick="location.href='/service/Q&A'">Q&A</a></div>
+</div>
